webscrap_project/SkinProductA/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
psycopg2-binary==2.9.10
lxml==5.3.0
elasticsearch==7.10.0

